// client/src/app/shared/models/product.ts
export type Product = {
  id: number;
  name: string;
  description: string;
  brand: string;
  category: string;
  subCategory: string;
  articleType: string;
  gender: string;
  ageGroup: string;
  baseColor: string;
  season: string;
  usage: string;
  material: string;
  pattern: string;
  fit: string;
  styleType: string;
  fashionType: string;
  year: number;
  price: number;
  discountPercentage: number;
  discountedPrice: number;
  rating: number;
  imageUrl: string;
  frontImageUrl: string;
  backImageUrl: string;
  searchImageUrl: string;
  tags: string[];
  isFeatured: boolean;
  isNewArrival: boolean;
  quantityInStock: number;
  createdAt: string;
  updatedAt: string;
}